// src/generated/whatsapp_event_service.ts
export interface CatchUpEventsComplete {
  headSequence: number;
}

export interface CatchUpEventsResponse {
  sequence?: number | undefined;
  messageChanged?: MessageChangeEvent | undefined;
  pollChanged?: PollChangeEvent | undefined;
  complete?: CatchUpEventsComplete | undefined;
  heartbeat?: Heartbeat | undefined;
}
// ...
export const CatchUpEventsComplete: MessageFns<CatchUpEventsComplete> = {
  encode(message: CatchUpEventsComplete, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.headSequence !== 0) {
      writer.uint32(8).uint64(message.headSequence);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CatchUpEventsComplete {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCatchUpEventsComplete();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.headSequence = longToNumber(reader.uint64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CatchUpEventsComplete {
    return {
      headSequence: isSet(object.headSequence)
        ? globalThis.Number(object.headSequence)
        : isSet(object.head_sequence)
        ? globalThis.Number(object.head_sequence)
        : 0,
    };
  },

  toJSON(message: CatchUpEventsComplete): unknown {
    const obj: any = {};
    if (message.headSequence !== 0) {
      obj.headSequence = Math.round(message.headSequence);
    }
    return obj;
  },

  create(base?: DeepPartial<CatchUpEventsComplete>): CatchUpEventsComplete {
    return CatchUpEventsComplete.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<CatchUpEventsComplete>): CatchUpEventsComplete {
    const message = createBaseCatchUpEventsComplete();
    message.headSequence = object.headSequence ?? 0;
    return message;
  },
};

function createBaseCatchUpEventsResponse(): CatchUpEventsResponse {
  return {
    sequence: undefined,
    messageChanged: undefined,
    pollChanged: undefined,
    complete: undefined,
    heartbeat: undefined,
  };
}

export const CatchUpEventsResponse: MessageFns<CatchUpEventsResponse> = {
  encode(message: CatchUpEventsResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.sequence !== undefined) {
      writer.uint32(8).uint64(message.sequence);
    }
    if (message.messageChanged !== undefined) {
      MessageChangeEvent.encode(message.messageChanged, writer.uint32(82).fork()).join();
    }
    if (message.pollChanged !== undefined) {
      PollChangeEvent.encode(message.pollChanged, writer.uint32(98).fork()).join();
    }
    if (message.complete !== undefined) {
      CatchUpEventsComplete.encode(message.complete, writer.uint32(162).fork()).join();
    }
    if (message.heartbeat !== undefined) {
      Heartbeat.encode(message.heartbeat, writer.uint32(794).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CatchUpEventsResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCatchUpEventsResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.sequence = longToNumber(reader.uint64());
          continue;
        }
        case 10: {
          if (tag !== 82) {
            break;
          }

          message.messageChanged = MessageChangeEvent.decode(reader, reader.uint32());
          continue;
        }
        case 12: {
          if (tag !== 98) {
            break;
          }

          message.pollChanged = PollChangeEvent.decode(reader, reader.uint32());
          continue;
        }
        case 20: {
          if (tag !== 162) {
            break;
          }

          message.complete = CatchUpEventsComplete.decode(reader, reader.uint32());
          continue;
        }
        case 99: {
          if (tag !== 794) {
            break;
          }

          message.heartbeat = Heartbeat.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CatchUpEventsResponse {
    return {
      sequence: isSet(object.sequence) ? globalThis.Number(object.sequence) : undefined,
      messageChanged: isSet(object.messageChanged)
        ? MessageChangeEvent.fromJSON(object.messageChanged)
        : isSet(object.message_changed)
        ? MessageChangeEvent.fromJSON(object.message_changed)
        : undefined,
      pollChanged: isSet(object.pollChanged)
        ? PollChangeEvent.fromJSON(object.pollChanged)
        : isSet(object.poll_changed)
        ? PollChangeEvent.fromJSON(object.poll_changed)
        : undefined,
      complete: isSet(object.complete) ? CatchUpEventsComplete.fromJSON(object.complete) : undefined,
      heartbeat: isSet(object.heartbeat) ? Heartbeat.fromJSON(object.heartbeat) : undefined,
    };
  },

  toJSON(message: CatchUpEventsResponse): unknown {
    const obj: any = {};
    if (message.sequence !== undefined) {
      obj.sequence = Math.round(message.sequence);
    }
    if (message.messageChanged !== undefined) {
      obj.messageChanged = MessageChangeEvent.toJSON(message.messageChanged);
    }
    if (message.pollChanged !== undefined) {
      obj.pollChanged = PollChangeEvent.toJSON(message.pollChanged);
    }
    if (message.complete !== undefined) {
      obj.complete = CatchUpEventsComplete.toJSON(message.complete);
    }
    if (message.heartbeat !== undefined) {
      obj.heartbeat = Heartbeat.toJSON(message.heartbeat);
    }
    return obj;
  },

  create(base?: DeepPartial<CatchUpEventsResponse>): CatchUpEventsResponse {
    return CatchUpEventsResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<CatchUpEventsResponse>): CatchUpEventsResponse {
    const message = createBaseCatchUpEventsResponse();
    message.sequence = object.sequence ?? undefined;
    message.messageChanged = (object.messageChanged !== undefined && object.messageChanged !== null)
      ? MessageChangeEvent.fromPartial(object.messageChanged)
      : undefined;
    message.pollChanged = (object.pollChanged !== undefined && object.pollChanged !== null)
      ? PollChangeEvent.fromPartial(object.pollChanged)
      : undefined;
    message.complete = (object.complete !== undefined && object.complete !== null)
      ? CatchUpEventsComplete.fromPartial(object.complete)
      : undefined;
    message.heartbeat = (object.heartbeat !== undefined && object.heartbeat !== null)
      ? Heartbeat.fromPartial(object.heartbeat)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}